feat: combo interface
