+++ src/types/index.ts
@@ -1,10 +1,10 @@
 import GraphEvent from '@antv/g-base/lib/event/graph-event';
-import { BBox } from '@antv/g-base/lib/types';
+import { BBox, AnimateCfg } from '@antv/g-base/lib/types';
 import Canvas from '@antv/g-canvas/lib/canvas';
 import ShapeBase from '@antv/g-canvas/lib/shape/base';
 import Node from '../item/node';
 import { IGraph } from '../interface/graph';
-import { IEdge, INode } from '../interface/item';
+import { IEdge, INode, ICombo } from '../interface/item';
 import { ILabelConfig } from '../interface/shape';
 
 // Math types
@@ -90,6 +90,190 @@ export type LoopConfig = Partial<{
   clockwise: boolean;
 }>;
 
+export interface LayoutConfig {
+  type?: string;
+  [key: string]: unknown;
+}
+
+
+export interface GraphAnimateConfig extends AnimateCfg {
+  /**
+   * 回调函数，用于自定义节点运动路径。
+   */
+  onFrame?: (item: Item, ratio: number, data?: GraphData, originAttrs?: ShapeStyle) => unknown;
+}
+
+export interface IModeOption {
+  type: string;
+  delegate?: boolean;
+  delegateStyle?: object;
+  updateEdge?: boolean;
+  trigger?: string;
+  enableDelegate?: boolean;
+  maxZoom?: number;
+  minZoom?: number;
+  enableOptimize?: boolean;
+  optimizeZoom?: number;
+  multiple?: boolean;
+  selectedState?: string;
+  includeEdges?: boolean;
+  direction?: 'x' | 'y';
+  shouldUpdate?: (e: IG6GraphEvent) => boolean;
+  shouldBegin?: (e: IG6GraphEvent) => boolean;
+  shouldEnd?: (e: IG6GraphEvent) => boolean;
+  onChange?: (item?: Item, judge?: boolean) => unknown;
+  onSelect?: (selectedNodes?: Item[], selectedEdges?: Item[]) => unknown;
+  onDeselect?: (selectedNodes?: Item[], selectedEdges?: Item[]) => unknown;
+  formatText?: (data: { [key: string]: unknown }) => string;
+}
+
+export type IModeType = string | IModeOption;
+
+export interface IMode {
+  default?: IModeType[];
+  [key: string]: IModeType[] | undefined;
+}
+
+// Graph 配置项中 state 的类型
+export interface IStates {
+  [key: string]: INode[];
+}
+
+export interface GraphOptions {
+  /**
+   * 图的 DOM 容器，可以传入该 DOM 的 id 或者直接传入容器的 HTML 节点对象
+   */
+  container: string | HTMLElement;
+  /**
+   * 指定画布宽度，单位为 'px'
+   */
+  width: number;
+  /**
+   * 指定画布高度，单位为 'px'
+   */
+  height: number;
+  /**
+   * renderer canvas or svg
+   */
+  renderer?: string,
+
+  fitView?: boolean;
+
+  layout?: LayoutConfig;
+
+  /**
+   * 图适应画布时，指定四周的留白。
+   * 可以是一个值, 例如：fitViewPadding: 20
+   * 也可以是一个数组，例如：fitViewPadding: [20, 40, 50,20]
+   * 当指定一个值时，四边的边距都相等，当指定数组时，数组内数值依次对应 上，右，下，左四边的边距。
+   */
+  fitViewPadding?: Padding;
+  /**
+   * 各种元素是否在一个分组内，决定节点和边的层级问题，默认情况下所有的节点在一个分组中，所有的边在一个分组中，当这个参数为 false 时，节点和边的层级根据生成的顺序确定。
+   * 默认值：true
+   */
+  groupByTypes?: boolean;
+
+  // 是否有向图
+  directed?: boolean;
+
+  groupStyle?: {
+    style?: {
+      [key: string]: ShapeStyle;
+    };
+  };
+
+  /**
+   * 当图中元素更新，或视口变换时，是否自动重绘。建议在批量操作节点时关闭，以提高性能，完成批量操作后再打开，参见后面的 setAutoPaint() 方法。
+   * 默认值：true
+   */
+  autoPaint?: boolean;
+
+  /**
+   * 设置画布的模式。详情可见G6中的Mode文档。
+   */
+  modes?: IMode;
+
+  /**
+   * 默认状态下节点的配置，比如 type, size, color。会被写入的 data 覆盖。
+   */
+  defaultNode?: {
+    shape?: string;
+    type?: string;
+    size?: number | number[];
+    color?: string;
+  } & ModelStyle;
+
+  /**
+   * 默认状态下边的配置，比如 type, size, color。会被写入的 data 覆盖。
+   */
+  defaultEdge?: {
+    shape?: string;
+    type?: string;
+    size?: number | number[];
+    color?: string;
+  } & ModelStyle;
+
+  /**
+   * Combo 默认配置
+   */
+  defaultCombo?: Partial<{
+    type: string;
+    size: number | number[];
+    color: string;
+  }> & ModelStyle;
+
+  nodeStateStyles?: { 
+    [key: string]: ShapeStyle | {
+      [key: string]: ShapeStyle
+    }
+  };
+
+  edgeStateStyles?: { 
+    [key: string]: ShapeStyle | {
+      [key: string]: ShapeStyle
+    }
+  };
+
+  // Combo 状态样式
+  comboStateStyles?: {
+    [key: string]: ShapeStyle | {
+      [key: string]: ShapeStyle
+    }
+  }
+
+  /**
+   * 向 graph 注册插件。插件机制请见：plugin
+   */
+  plugins?: any[];
+  /**
+   * 是否启用全局动画。
+   */
+  animate?: boolean;
+
+  /**
+   * 动画配置项，仅在animate为true时有效。
+   */
+  animateCfg?: GraphAnimateConfig;
+  /**
+   * 最小缩放比例
+   * 默认值 0.2
+   */
+  minZoom?: number;
+  /**
+   * 最大缩放比例
+   * 默认值 10
+   */
+  maxZoom?: number;
+
+  groupType?: string;
+
+  /**
+   * Edge 是否连接到节点中间
+   */
+  linkCenter?: boolean;
+}
+
 // model types (node edge group)
 export type ModelStyle = Partial<{
   [key: string]: unknown;
@@ -101,7 +285,7 @@ export type ModelStyle = Partial<{
   };
   // loop edge config
   loopCfg: LoopConfig;
-  labelCfg?: ILabelConfig;
+  labelCfg: ILabelConfig;
 }>;
 
 export type LabelStyle = Partial<{
@@ -227,6 +411,21 @@ export interface NodeConfig extends ModelConfig {
   description?: string;
 }
 
+export interface ComboConfig {
+  id: string;
+  parentId?: string;
+  // Combo 类型，默认 rect，值为定义的 combo 的名称
+  type: string;
+  // Combo 标题
+  title: string | LabelStyle;
+  style: ShapeStyle;
+  stateStyles: {
+    [key: string]: ShapeStyle | {
+      [key: string]: ShapeStyle
+    }
+  };
+}
+
 export interface EdgeConfig extends ModelConfig {
   id?: string;
   source?: string;
@@ -273,6 +472,7 @@ export interface GraphData {
   nodes?: NodeConfig[];
   edges?: EdgeConfig[];
   groups?: GroupConfig[];
+  combos?: ComboConfig[];
 }
 
 export interface TreeGraphData {
@@ -365,10 +565,10 @@ export interface IG6GraphEvent extends GraphEvent {
   target: Item & Canvas;
 }
 
-// Node Edge 实例或ID
-export type Item = INode | IEdge;
+// Node Edge Combo 实例
+export type Item = INode | IEdge | ICombo;
 
-export type ITEM_TYPE = 'node' | 'edge' | 'group';
+export type ITEM_TYPE = 'node' | 'edge' | 'combo' | 'group';
 
 export type NodeIdxMap = {
   [key: string]: number;
@@ -382,9 +582,4 @@ export interface ViewPortEventParam {
 
 export interface Indexable<T> {
   [key: string]: T;
-}
-
-export interface LayoutConfig {
-  type?: string;
-  [key: string]: unknown;
 }
