+++ tests/unit/graph/controller/mode-spec.ts
@@ -1,6 +1,6 @@
 import { ModeController } from '../../../../src/graph/controller';
 import Graph from '../../../../src/graph/graph';
-import { GraphOptions, IGraph, IModeOption } from '../../../../src/interface/graph';
+import { GraphOptions, IModeOption } from '../../../../src/types'
 
 const div = document.createElement('div');
 div.id = 'graph-spec';
